BRD-016: Adding more value types

+++ templates/noc/lib/compiler/op-val.php
@@ -12,12 +12,12 @@ class OpVal implements Compilable {
             'equals' => array(
                 'boolean' => true,
                 'integer' => true,
-                'double' => true,
+                'float' => true,
                 'string' => true
             ),
             'lessThan' => array(
                 'integer' => true,
-                'double' => true
+                'float' => true
             )
         );
     }
